feat(faq): abstract sidebar markup component
